+++ redshell/src/app/chat.rs
@@ -291,20 +291,20 @@ mod tests {
     #[allow(unused)]
     const DOWN: Action = Action::KeyPress { key: Key::Down };
 
-    #[coverage_helper::test]
+    #[test]
     fn test_submit_reply() {
         let mut app = app_dm(&["hello"], 0);
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "hello")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_select_submit() {
         let mut app = app_dm(&["hello", "goodbye"], 0);
         assert_input!(app.input(RIGHT) taints, .is_empty());
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "goodbye")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_select_return_submit() {
         let mut app = app_dm(&["hello", "goodbye"], 0);
         assert_input!(app.input(RIGHT) taints, .is_empty());
@@ -312,7 +312,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "hello")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_select_max_right() {
         let mut app = app_dm(&["hello", "goodbye"], 0);
         assert_input!(app.input(RIGHT) taints, .is_empty());
@@ -321,7 +321,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "goodbye")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_select_max_left() {
         let mut app = app_dm(&["hello", "goodbye"], 1);
         assert_input!(app.input(LEFT) clean, .is_empty());
@@ -330,7 +330,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "hello")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_receive_option() {
         let mut app = app_dm(&[], 0);
         app.on_event(
@@ -340,7 +340,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "hi")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_receive_next_option() {
         let mut app = app_dm(&[], 0);
         app.on_event(
@@ -351,7 +351,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "no")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_switch_dms() {
         let mut app = app_dm(&["normal", "human", "words"], 0);
         app.on_event(
@@ -362,7 +362,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("meowza", "hi")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_switch_dms_extra() {
         let mut app = app_dm(&["normal", "human", "words"], 0);
         app.on_event(
@@ -375,7 +375,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("meowza", "hi")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_switch_dms_back() {
         let mut app = app_dm(&["normal", "human", "words"], 0);
         app.on_event(
@@ -387,7 +387,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "normal")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_switch_dms_back_extra() {
         let mut app = app_dm(&["normal", "human", "words"], 0);
         app.on_event(
@@ -401,7 +401,7 @@ mod tests {
         assert_input!(app.input(ENTER) taints, == &[Event::player_chat("targette", "normal")]);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_add_notifs() {
         let mut app = app_dm(&[], 0);
         app.on_event(&Event::npc_chat("targette", "hi", &[]), false);
@@ -410,7 +410,7 @@ mod tests {
         assert_eq!(app.notifs(), 3);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_clear_notifs() {
         let mut app = app_dm(&[], 0);
         app.on_event(&Event::npc_chat("targette", "hi", &[]), false);
@@ -419,7 +419,7 @@ mod tests {
         assert_eq!(app.notifs(), 0);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_add_notifs_more_dms() {
         let mut app = app_dm(&[], 0);
         app.on_event(&Event::npc_chat("targette", "hi", &[]), false);
@@ -437,7 +437,7 @@ mod tests {
         assert_eq!(app.notifs(), 6);
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn test_clear_notifs_more_dms_other_side() {
         let mut app = app_dm(&[], 0);
         app.on_event(&Event::npc_chat("targette", "hi", &[]), false);
