+++ tuig/Cargo.toml
@@ -1,0 +1,47 @@
+[package]
+name = "tuig"
+version = "0.0.1"
+edition = "2021"
+
+[features]
+# See tuig::docs::features for crucial information on selecting them
+__io = []
+io_nop = ["__io"]
+io_cli = ["__io", "crossterm"]
+__io_gui = ["__io", "winit"]
+io_gui_softbuffer = ["__io_gui", "softbuffer", "fontdue", "rayon"]
+__run = []
+run_orig = ["__run"]
+run_single = ["__run"]
+run_rayon = ["__run", "rayon", "crossbeam"]
+
+plat_ssh = [
+    "io_cli",
+    "run_rayon", "run_single",
+]
+plat_steam = [
+    "io_gui_softbuffer",
+    "run_rayon", "run_single",
+]
+# plat_floppy = [
+#     "io_cli",
+#     "run_single",
+# ]
+# plat_wasm = [
+#     "io_web",
+#     "run_single",
+# ]
+
+[dependencies]
+paste = "1.0.5"
+rand = { version = "0.8.5", features = ["small_rng"] }
+
+crossterm = { version = "0.20.0", optional = true }
+winit = { version = "0.27.5", optional = true }
+softbuffer = { version = "0.1.1", optional = true }
+fontdue = { version = "0.7.2", optional = true }
+rayon = { version = "1.6.1", optional = true }
+crossbeam = { version = "0.8.2", optional = true }
+
+[dev-dependencies]
+mock_instant = "0.2.1"
