+++ tuig/src/agent.rs
@@ -1,3 +1,5 @@
+use crate::game::{Message, Replies};
+
 use core::fmt;
 use std::{
     sync::{
@@ -110,17 +112,17 @@ mod cf_test {
 
     use super::{ControlFlow, Instant};
 
-    #[coverage_helper::test]
+    #[test]
     fn continue_ready() {
         assert!(ControlFlow::Continue.is_ready())
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn kill_unready() {
         assert!(!ControlFlow::Kill.is_ready());
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn wait_handle_readies_after_touch() {
         let (cf, wh) = ControlFlow::wait();
         assert!(!cf.is_ready());
@@ -128,7 +130,7 @@ mod cf_test {
         assert!(cf.is_ready());
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn sleep_until_readies_after_time() {
         let cf = ControlFlow::sleep_until(Instant::now() + Duration::from_millis(100));
         assert!(!cf.is_ready());
@@ -138,7 +140,7 @@ mod cf_test {
         assert!(cf.is_ready());
     }
 
-    #[coverage_helper::test]
+    #[test]
     fn sleep_for_readies_after_time() {
         let cf = ControlFlow::sleep_for(Duration::from_millis(100));
         assert!(!cf.is_ready());
@@ -146,5 +148,36 @@ mod cf_test {
         assert!(!cf.is_ready());
         MockClock::advance(Duration::from_millis(60));
         assert!(cf.is_ready());
+    }
+}
+
+/// An agent in the system, which can react to events.
+///
+/// Events are processed in 'rounds'. There's a list of 'current' events, which are fed into every actor at the same
+/// time. Then all of the replies are collected, and those are the 'current' events for the next round.
+///
+/// As that implies, events are inherently ephemeral -- none persist more than one round.
+pub trait Agent<M: Message>: Send + Sync {
+    /// Called once on (re)start, to queue any starting events/ControlFlow as necessary. This will always be called
+    /// before `react`.
+    ///
+    /// By default, does nothing and returns [`ControlFlow::Continue`] to allow [`Self::react`] to be called, under
+    /// the assumption that your interesting code sits there.
+    #[cfg_attr(coverage, no_coverage)]
+    fn start(&mut self, _replies: &mut Replies<M>) -> ControlFlow {
+        ControlFlow::Continue
+    }
+
+    /// React to the events of a round, indicating when the agent should be called next and optionally queueing some
+    /// more events.
+    ///
+    /// Limitations on the [`Extend`] trait mean we just use the concrete type `Vec`. **Do not** do anything except
+    /// pushing/extending/otherwise adding elements.
+    ///
+    /// By default, does nothing and returns [`ControlFlow::Kill`], under the assumption that you'd have implemented
+    /// `react` if you wanted your agent to stay alive and do things.
+    #[cfg_attr(coverage, no_coverage)]
+    fn react(&mut self, _event: &M, _replies: &mut Replies<M>) -> ControlFlow {
+        ControlFlow::Kill
     }
 }
