+++ tuig/src/io/mod.rs
@@ -1,0 +1,9 @@
+//! Common code and types between input and output.
+
+pub mod clifmt;
+pub mod helpers;
+pub mod input;
+pub mod output;
+pub mod sys;
+pub mod widgets;
+pub mod xy;
