+++ tuig/src/lib.rs
@@ -1,0 +1,13 @@
+mod game;
+mod agent;
+mod io;
+mod util;
+
+/// Re-exports you may find useful when using tui-engine; use it like:
+/// 
+/// ```rs
+/// use tuig::prelude::*;
+/// ```
+pub mod prelude {
+
+}
